Introduced a new class PHPClassDef as discussed in PR #73, and reverted ClassDefStatement to its previous state.
